+++ imio/dms/mail/content/member_area.py
@@ -1,12 +1,12 @@
 # -*- coding: utf-8 -*-
 from imio.dms.mail.interfaces import IMemberAreaFolder
 from plone.dexterity.content import Container
-from zope.interface import implements
+from zope.interface import implementer
 
 
+@implementer(IMemberAreaFolder)
 class MemberArea(Container):
     """
     MemberArea class
     """
 
-    implements(IMemberAreaFolder)