+++ imio/dms/mail/testing.py
@@ -1,6 +1,9 @@
 # -*- coding: utf-8 -*-
 from collections import OrderedDict
+from collective.contact.plonegroup.config import get_registry_functions
 from collective.contact.plonegroup.config import get_registry_organizations
+from collective.contact.plonegroup.config import set_registry_functions
+from collective.iconifiedcategory.utils import calculate_category_id
 from imio.dms.mail.utils import set_dms_config
 from imio.dms.mail.utils import sub_create
 from imio.helpers.cache import setup_ram_cache
@@ -8,6 +11,7 @@ from imio.helpers.content import get_object
 from imio.helpers.workflow import do_transitions
 from imio.pyutils.system import runCommand
 from itertools import cycle
+from pathlib import Path
 from plone import api
 from plone.api.validation import at_least_one_of
 from plone.app.robotframework.remote import RemoteLibrary
@@ -340,6 +344,57 @@ def change_user(portal, user="siteadmin"):
     login(portal, user)
 
 
+def ensure_demand_sign(portal, org_uid, userids=("chef", "dirg")):
+    """Register the 'demand_sign' plonegroup function and add users to '<org_uid>_demand_sign'."""
+    functions = get_registry_functions()
+    if u"demand_sign" not in [fct["fct_id"] for fct in functions]:
+        functions.append(
+            {"fct_title": u"Demande Sign.", "fct_id": u"demand_sign", "fct_orgs": [],
+             "fct_management": False, "enabled": True}
+        )
+        set_registry_functions(functions)
+    for userid in userids:
+        portal.acl_users.source_groups.addPrincipalToGroup(userid, "%s_demand_sign" % org_uid)
+
+
+def create_sign_request(portal, oid="sr1", signers=None, nb_files=1, org_uid=None,
+                        assigned_user="chef", esign=False, title=u"Demande de signature test"):
+    """Create a sign_request in the 'requests' folder with optional signers and appendix files.
+
+    Returns a (request, files) tuple.
+    """
+    if org_uid is None:
+        org_uid = get_registry_organizations()[0]
+    ensure_demand_sign(portal, org_uid)
+    params = {
+        "title": title,
+        "treating_groups": org_uid,
+        "assigned_user": assigned_user,
+        "recipient_groups": [],
+        "signers": signers,
+        "esign": esign,
+    }
+    request = sub_create(portal["requests"], "sign_request", datetime.datetime.now(), oid, **params)
+    ct = portal["annexes_types"]["sign_request_appendix_files"]["sign-request-appendix-file"]
+    filename = u"3-degradation-voirie.odt"
+    filepath = "%s/batchimport/toprocess/requests/%s" % (imio.dms.mail.__path__[0], filename)
+    files = []
+    for i in range(nb_files):
+        with open(filepath, "rb") as fo:
+            file_object = NamedBlobFile(fo.read(), filename=filename)
+            files.append(
+                createContentInContainer(
+                    request,
+                    "dmsappendixfile",
+                    id="file%s" % i,
+                    scan_id="012999900000601",
+                    file=file_object,
+                    content_category=calculate_category_id(ct),
+                )
+            )
+    return request, files
+
+
 @timecall
 def create_im_mails(tc, start=1, end=100, senders=[], transitions=[], by_days=200):
     """Create a number of im"""
@@ -348,7 +403,7 @@ def create_im_mails(tc, start=1, end=100, senders=[], transitions=[], by_days=20
 
     filespath = "%s/batchimport/toprocess/incoming-mail" % imiodmsmail.__path__[0]
     files = [safe_unicode(name) for name in os.listdir(filespath)
-             if os.path.splitext(name)[1][1:] in ("pdf", "doc", "jpg")]
+             if Path(name).suffix[1:] in ("pdf", "doc", "jpg")]
     files_cycle = cycle(files)
 
     # intids = getUtility(IIntIds)